requirements.txt
# Commonly used
pre-commit
pytest
pandas
numpy
black
ruff

# FastAPI requirement
fastapi~=0.111.0
uvicorn[standard]~=0.29.0

# Flask requirement
flask~=3.0.3

# Ngrok requirements
pyngrok

python-dotenv~=1.0.1
PyYAML~=6.0.1
